Format multiple authors
Cargo's `CARGO_PKG_AUTHORS` env var separates author names with a colon,
which is not very pretty to look at. So, let's replace it with a comma.

Clap has the same functionality it its `crate_authors!` macro but as
it's only a few lines of code I didn't want to call that.